Add "Redirect Back to Origin Page" setting
This setting would override the default action of redirecting the user
to the home page after a successful login and instead redirect the user
back to the page on which they clicked the OpenID Connect login button.
This would cause the login process to proceed in a traditional WordPress
fashion.