Store states as their own option
We found that there's a potential race-condition with storing an array
array of valid states in one option. One process could potentially write
an old version of an array to the database after a different one had
already updated it, causing the old version to be the current one.

By storing each state as it's own option we can ensure that the correct
data is being stored and will be retrieved successfully.

One negative side-effect that we will need to address is the clean-up of
the state value belonging to aborted expired logins. ie. if a user
attempts to sign in to the intranet (we set the state) then forgets
their password and leaves the state will still exist in the database.
We will need a separate job that cleans up expired states on a regular
basis.